split metaballs into Sketch and Mesh for 2D and 3D respectively
